Fixed a lot of false positives

+++ Helper/ObjectMetadata.cs
@@ -28,14 +28,14 @@
         /// When the value is below 0 it means a hyper is generated
         /// Otherwise it must be a dash or walk
         /// </summary>
-        public int DistanceToHyper { get; set; }
+        public float DistanceToHyper { get; set; }
 
         /// <summary>
         /// The amount of extra distance needed until a Dash movement is required
         /// When the value ie below 0 it means a dash movement is mandatory
         /// Otherwise it must be a walk
         /// </summary>
-        public int DistanceToDash { get; set; }
+        public float DistanceToDash { get; set; }
 
         public MovementType MovementType { get; set; }
     }
